Catching the bug in TrackingFunctionPF.hpp
This revision will compile, but won't run correctly. I'm trying to catch
the bug in TrackingFunctionPF.hpp and it seems that it is related to
BGR->HSV conversion. If someone is interested in giving me a hand, the
problem is in TrackingFunctionPF.hpp, the constructor of
TrackingFunctionPF::TrackingHistogram. After HSV conversion done, many
pixels on HSV picture have abnormal values.